[feat] 환자 기본정보 조회& 조치 방법 저장 api구현 및 스웨거 설정 변경

+++ src/main/java/AIYA/com/FAIN/repository/ReportRepository.java
@@ -3,6 +3,7 @@ package AIYA.com.FAIN.repository;
 import AIYA.com.FAIN.entity.Reports;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportRepository extends JpaRepository<Reports,Integer> {
+public interface ReportRepository extends JpaRepository<Reports,Long> {
+
 
 }
